requirements.txt
pyspark==4.0.1
beautifulsoup4==4.13.5
pandas==2.3.2
requests==2.32.5
pymongo==4.14.1
python-dotenv==1.1.1
